challenge-polyai-smartsort/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pymongo==4.6.0
python-dotenv==1.0.0
Pillow==10.3.0
numpy==1.26.4
tensorflow==2.16.1
scikit-learn==1.4.2
pandas==2.2.2
pydantic==2.5.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1